Fix remaining compiler warnings in rendering, models, console, widgets
stb_include.cpp:175 [-Wformat]:
  next_line_after is int64_t but format used %zu (size_t). Fix: %lld.

model.cpp:135 [-Winvalid-offsetof]:
  offsetof on FSpriteModelFrame (non-standard-layout). Well-defined on
  all target compilers; suppress locally with pragma clang diagnostic.

c_buttons.cpp:160 [-Wsign-compare]:
  int loop counter vs Buttons.Size() (unsigned int). Cast Size() to int.

netstartwindow.cpp:92 [-Wsign-compare]:
  size_t loop var vs maxpos (int). Cast maxpos to size_t.

+++ src/common/console/c_buttons.cpp
@@ -157,7 +157,7 @@ void ButtonMap::GetAxes ()
 	float joyaxes[NUM_AXIS_CODES];
 	I_GetAxes(joyaxes);
 
-	for (int i = 0; i < Buttons.Size(); i++)
+	for (int i = 0; i < (int)Buttons.Size(); i++)
 	{
 		FButtonStatus &btn = Buttons[i];
 		FString &btn_name = NumToName[i];
